+++ lib/utils/enums/picker_mode.dart
@@ -7,3 +7,9 @@ enum PickerType {
   jalali,
   dateTime,
 }
+
+// simple calender view or complex
+enum CalenderType {
+  simple,
+  complex,
+}